Added more methods to help towards capsulation
All settings are now stored within a hash.
Added method "get_mail", returning the mail with the specified id.
Added method "inbox_total", returning the total number of mails in an inbox.
Added method "configure", which updates user preferences in the database.
Added method "mark_as", which updates the "read" field for the messages with the specified message_ids or all of them, if :all is given instead of an array.

Fixed method "get_inbox".