Added various methods
Added ::attachment_names, which returns an array with all the file names attached to the mail with the specified mail id.
Added ::mail_has_attachments, which returns true or false, depending on whether the mail with the specified id has attachments.
Added ::del_user, which deletes a user from the database along with all his mails, attachments and user configuration.
Added ::del_mail, which deletes the mail with the specified id from the database.